+++ src/messagingsession/DefaultMessagingSession.ts
@@ -70,7 +70,7 @@ export default class DefaultMessagingSession implements MessagingSession {
 
   async startAsync(): Promise<void> {
     if (this.isClosed()) {
-      await this.startConnecting(false);
+      this.startConnecting(false);
     } else {
       this.logger.info('messaging session already started');
     }
@@ -111,10 +111,10 @@ export default class DefaultMessagingSession implements MessagingSession {
     });
   }
 
-  private async startConnecting(reconnecting: boolean) {
-    // reconnect needs to re-resolve endpoint url, which will also refresh credentials on client if they are expired
-    let endpointUrl = !reconnecting ? this.configuration.endpointUrl : null;
-    if (endpointUrl == null) {
+  private async startConnecting(reconnecting: boolean): Promise<void> {
+    // reconnect needs to re-resolve endpoint url, which will also refresh credentials on client if they are expired.
+    let endpointUrl = !reconnecting ? this.configuration.endpointUrl : undefined;
+    if (endpointUrl === undefined) {
       const endpoint = await this.configuration.chimeClient.getMessagingSessionEndpoint().promise();
       endpointUrl = endpoint.Endpoint.Url;
     }