+++ test/messagingsession/DefaultMessagingSession.test.ts
@@ -39,6 +39,8 @@ describe('DefaultMessagingSession', () => {
   let dommMockBehavior: DOMMockBehavior;
   let domMockBuilder: DOMMockBuilder;
 
+  let getMessSessionCnt = 0;
+
   const chimeClient = {
     config: {
       region: 'us-east-1',
@@ -50,7 +52,8 @@ describe('DefaultMessagingSession', () => {
     },
     getMessagingSessionEndpoint: function () {
       return {
-        promise : async function(): Promise<any> {
+        promise: async function (): Promise<any> {
+          getMessSessionCnt++;
           return {
             Endpoint: {
               Url: ENDPOINT_URL
@@ -90,13 +93,8 @@ describe('DefaultMessagingSession', () => {
     dommMockBehavior = new DOMMockBehavior();
     dommMockBehavior.webSocketSendEcho = true;
     domMockBuilder = new DOMMockBuilder(dommMockBehavior);
-    configuration = new MessagingSessionConfiguration(
-      'userArn',
-      '123',
-      null,
-      chimeClient,
-      {}
-    );
+    getMessSessionCnt = 0;
+    configuration = new MessagingSessionConfiguration('userArn', '123', undefined, chimeClient, {});
     configuration.reconnectTimeoutMs = 100;
     configuration.reconnectFixedWaitMs = 40;
     configuration.reconnectShortBackoffMs = 10;
@@ -137,6 +135,21 @@ describe('DefaultMessagingSession', () => {
     it('Can start', done => {
       messagingSession.addObserver({
         messagingSessionDidStart(): void {
+          expect(getMessSessionCnt).to.be.eq(1);
+          done();
+        },
+      });
+      messagingSession.start();
+      new TimeoutScheduler(10).start(() => {
+        webSocket.send(SESSION_SUBSCRIBED_MSG);
+      });
+    });
+
+    it('Can start with hardcoded config url', done => {
+      configuration.endpointUrl = ENDPOINT_URL;
+      messagingSession.addObserver({
+        messagingSessionDidStart(): void {
+          expect(getMessSessionCnt).to.be.eq(0);
           done();
         },
       });
@@ -277,6 +290,7 @@ describe('DefaultMessagingSession', () => {
         messagingSessionDidStop(_event: CloseEvent): void {
           expect(didStartConnecting).to.be.eq(2);
           expect(didStartCount).to.be.eq(1);
+          expect(getMessSessionCnt).to.be.eq(2);
           done();
         },
       });